@@ -1,6 +1,6 @@
 module go.opentelemetry.io/build-tools/chloggen
 
-go 1.17
+go 1.18
 
 require (
 	github.com/stretchr/testify v1.8.0
