@@ -1,23 +1,28 @@
 module go.opentelemetry.io/build-tools/crosslink
 
-go 1.16
+go 1.18
 
 require (
 	github.com/google/go-cmp v0.5.8
+	github.com/otiai10/copy v1.7.0
 	github.com/spf13/cobra v1.5.0
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.8.0
+	go.opentelemetry.io/build-tools v0.0.0-20220321164008-b8e03aff061a
 	go.uber.org/zap v1.23.0
+	golang.org/x/mod v0.5.1
 )
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/kr/pretty v0.2.0 // indirect
-	github.com/otiai10/copy v1.7.0
-	go.opentelemetry.io/build-tools v0.0.0-20220321164008-b8e03aff061a
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.7.0 // indirect
-	golang.org/x/mod v0.5.1
+	golang.org/x/xerrors v0.0.0-20191011141410-1b5146add898 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace go.opentelemetry.io/build-tools => ../
