@@ -1,8 +1,15 @@
 module go.opentelemetry.io/build-tools
 
-go 1.15
+go 1.18
 
 require (
 	github.com/stretchr/testify v1.8.0
 	golang.org/x/mod v0.5.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/xerrors v0.0.0-20191011141410-1b5146add898 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
