@@ -1,11 +1,20 @@
 module go.opentelemetry.io/build-tools/issuegenerator
 
-go 1.16
+go 1.18
 
 require (
 	github.com/google/go-github v17.0.0+incompatible
-	github.com/google/go-querystring v1.0.0 // indirect
 	github.com/joshdk/go-junit v0.0.0-20210226021600-6145f504ca0d
 	go.uber.org/zap v1.23.0
 	golang.org/x/oauth2 v0.0.0-20210220000619-9bb904979d93
+)
+
+require (
+	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/google/go-querystring v1.0.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
+	go.uber.org/multierr v1.6.0 // indirect
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	google.golang.org/appengine v1.6.6 // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
 )
