@@ -4,13 +4,14 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 package crosslink
 
 // Helper functions that are used by multiple test files
@@ -64,8 +65,8 @@ func createTempTestDir(testName string) (string, error) {
 		return "", fmt.Errorf("failed to make temp director: %w", err)
 	}
 
-	mockDataDir := filepath.Join(mockDataDir, testName)
-	err = cp.Copy(mockDataDir, tmpRootDir)
+	mockDir := filepath.Join(mockDataDir, testName)
+	err = cp.Copy(mockDir, tmpRootDir)
 	if err != nil {
 		return "", fmt.Errorf("failed to copy mock data into temp: %w", err)
 	}
