@@ -23,7 +23,7 @@ type ErrGitTagsAlreadyExist struct {
 	tagNames []string
 }
 
-func (e *ErrGitTagsAlreadyExist) Error() string {
+func (e ErrGitTagsAlreadyExist) Error() string {
 	return fmt.Sprintf("all git tags checked already exist:\n%s", strings.Join(e.tagNames, "\n"))
 }
 
@@ -31,7 +31,7 @@ type ErrInconsistentGitTagsExist struct {
 	tagNames []string
 }
 
-func (e *ErrInconsistentGitTagsExist) Error() string {
+func (e ErrInconsistentGitTagsExist) Error() string {
 	return fmt.Sprintf("git tags inconsistent for module set (some but not all tags in module set):\n%s", strings.Join(e.tagNames, "\n"))
 }
 
